Update README.md
Desktop is a subset of Transient running widgets in the Web browser with the Haste compiler. The widgets are first class, that means that there is an widget algebra using standard haskell combinators.

latest additions:

A monad for reactive programming part 2 describes the monadic reactive mechanism of Desktop
A Tutorial for creating client side applications using Haste and Desktop
ajax with de-inversion of control
WebSockets with de-inversion of control
A accounting application example that shows how to include a javascript library -the google chart library- as a widget.